+++ MyEssentialFeed/MyEssentialFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,25 @@
+//
+//  FeedCachePolicy.swift
+//  MyEssentialFeed
+//
+//  Created by Tiago Chaves on 02/01/24.
+//
+
+import Foundation
+
+internal final class FeedCachePolicy {
+    private init() {}
+    
+    private static let calendar = Calendar(identifier: .gregorian)
+    
+    private static var maxCacheAgeInDays: Int {
+        return 7
+    }
+    
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
